IO: eliminate race in close
I knew that I'll hit this bug anyway. Here we MUST flush twice as
flush at read suspend is not enough. If a writer closes then the first
flush does flush and it closes so all readers will see all pending data.
If a foreign thread does close then the first flush is makes no
difference however the second is required for proper cleanup and
to do tryTerminate