Replace ConcurrentHashMap by mutableMapOf
Because the maps are just a best effort cache, we don't need the
concurrency safety that ConcurrentHashMap offers.

If the scopes or jobs are accessed concurrently, there may be multiple
created instances, but eventually, an attempt to remove them from the
caching map will always be performed when the lifecycle is destroyed.