Replace arraycopy with stdlib copyInto extension
The intent is to simplify the code and improve readability, taking
advantage of the default argument values of copyInto that Java-like
arraycopy doesn't provide.

arraycopy implementations are no longer used now, so this commit also
removes them, plus its no longer needed test.

This also reduces the needed test surface of the library, relying on
stdlib implementation instead.

Besides the used import removal, the changes performed in the
NamedDispatchers.kt file are exactly the same as the ones performed in
ConflatedBroadcastChannel.kt file.