Get rid of AbstractCoroutine.defaultResumeMode
It was only needed to customize resume mode from within of
AbstractCoroutine.resumeWith method. This is now achieved by
a separate open fun afterResume. Also, afterCompletionInternal is
now renamed to afterCompletion and is only used from the job was
completed from unknown context and thus should be resumed in
a default way. This "default way" is now "cancellable resume"
(to minimize the amount of code duplication) which does not introduce
any functional difference, since that only happens on cancellation.