Fix Byte Buddy use in debug jar
The `shadow` configuration extends `api` now, which means that all
the `api`-style dependencies of this module are not bundled in the
fat jar and are instead mentioned in the POM file as `RUNTIME`
dependencies.

This means that now JNA is not included with the jar, but when
someone depends on the jar to later call `DebugProbes.install()`,
JNA is fetched as a transitive dependency.

This is what we want: JNA is not needed in the jar, because when
the debugger is run as a Java agent, attachment of Byte Buddy
happens in `Installer`, and JNA-based attach emulation is not used;
Byte Buddy is also written is such a way that if JNA is not in the
class path, nothing breaks unless one tries to use attach
emulation.