Fix publication validator not looking in the local maven repository
During the change of the publication validator, a bug was
introduced that led to MavenPublicationValidator being run not on
the artifacts from the local Maven repository but on classfiles
from the corresponding subproject. This is a problem because this
test is for the behavior of the atomicfu plugin, which could in
theory produce nice classfiles in one place but wrong ones in the
other, and the only important thing to test is whether the
published classfiles are good.

Now, this is fixed.