Release coroutine block after lazy start
When starting a coroutine lazily by using CoroutineStart.LAZY, it
creates a Continuation that would launch the coroutine block, which is
stored in the returned Job or Deferred to be invoked lazily. However,
this reference is never cleared, even after the coroutine ends.

This prevents the coroutine block and it's enclosing class instance from
being garbage collected if the Job or Deferred is retained in another
class, or while child coroutines are running (as they also hold a
reference to the Job).

Fixed by clearing the reference to the Continuation after invoking it.