Handle undeliverable errors in rxObservable
If an error can't be delivered to the observer due to the observer
already having disposed the subscription, then it should be reported
to the RX global exception handler. This was already being done in
the implementations for rxSingle, rxMaybe, and rxCompletable, but
not in rxObservable.

Copied the test over from the other implementations as well.

Also removed the previous test in ObservableTest, which actually was
testing rxSingle instead of rxObservable, and was too complicated
and dependent on race conditions (both rxSingle and rxObservable now
have a simple and deterministic test for undeliverable error
handling). It was also not well implemented, and I originally tried
to fix it to be more robust in commit 58ad0cdfca81ab2d5c5c410f889293c51beede75.

Fixes #2173.