Bug fix and better LockFreeLinkedList verification with Lincheck
Fixes a rare crash of LockFreeLinkedListAtomicLFStressTest, e.g:
https://teamcity.jetbrains.com/buildConfiguration/KotlinTools_KotlinxCoroutines_NightlyStressWindows/3215239
with AssertionError at kotlinx.coroutines.internal.LockFreeLinkedListNode.validateNode

This was happening because during a concurrent execution of complex atomic operation a thread that add a node to the list could be helped to complete by other threads and the newly added node could be already removed. Now, the original adder thread will execute its add operation too, adding the removed element back to the list, leaving the list in state that is logically correct, but which is not correct physically (with a removed node still in the list).

The Lincheck test that consistently reproduces this problem is added and the fix is introduced.