Fix the coroutine debugger in the IDE
https://github.com/Kotlin/kotlinx.coroutines/pull/2461 introduced
a problem where the coroutine debugger in the IDE would crash:
https://github.com/Kotlin/kotlinx.coroutines/issues/2690

According to @nikita-nazarov, the coroutine debugger crashes at a
point where a stack trace is already available and the IDE plugin
attempts to look up every class listed in the stack trace.

The new classes listed in the artificial stack frames were never
loaded, as only their names were accessed. Now, we force these
classes to be loaded, and it seems like the crash disappears.