docs: improve Children of a coroutine example
I think flipping the order of the print and delay makes it more clear that one coroutine is canceled while the other lives on.  Also I added a note that explains why the delay is necessary.  The need for the second delay wasn't immediately clear to me on first read of the coroutine guide.