Fix a suppressed warning
This also fixes the build with the in-development 1.8.0 Kotlin
compiler.

Before, when constructing `Maybe<A & Any>`, we had `T = A?`.
This way, `MaybeEmitter<T & Any>` meant `MaybeEmitter<A? & Any>`.
However, with the new version, type inference must have changed,
so it came to be that `T = A`, breaking the code.