Stabilize Channel.invokeOnClose
CompletionHandler is not user deliberately, as its contract requires some additional refinement along with `onCancelling` handler stabilization. Note that replacing functional type with the very same typealias is backwards-compatible in the current state of linkage, so we are not giving up any future opportunities.

Also, fix behavioral mismatch: `CancellationException` is supplied (and always has been) to `invokeOnClose` when a channel was cancelled normally instead of `null` as documentation stated.
This behaviour is aligned with other cancellation handlers and also allows handler to distinguish whether the channel was closed or cancelled.

Fixes #3358