Fix `limitedParallelism` doing dispatches when it has no tasks (#3672)
To support fairness w.r.t. the tasks that want to run on the
wrapped dispatcher outside of its view, `LimitedDispatcher`
redispatches the work loop periodically. This can happen even when
there are no more tasks in `LimitedDispatcher`'s queue.
Typically, this behavior is completely benign: after being
redistpatched, the new runner just checks the queue, sees that
there is nothing there, and leaves. However, with closable
dispatchers, this affects correctness, as this redispatch may
happen even after all the dispatched tasks were run and the
dispatcher was closed.