Update kotlinx-coroutines-core build file for IDE friendliness (#3795)
The build configuration modification was done to prevent unexpected "shared jvm" behavior in IDE due to the improper jvmCoreMain to jvmMain dependency. A conditional dependency inclusion is added, which will only engage in CLI builds and not in the IDE. This change addresses the issue KTIJ-25959 and avoids IDE confusion with multi-compilation scenarios.