Add native targets to the smokeTest project in integration tests.
As there were no native targets in smokeTest, the problem with lacking atomicfu dependency (KT-64111) was not addressed in time. The bug was revealed after  release of kotlinx.coroutines with atomicfu-gradle-plugin 0.23.1. Coroutines enabled Native IR transformation, and for this mode atomicfu-gradle-plugin only provided compile dependency to atomicfu, though implementation dependency was still necessary.