Change the order of atomicfu and kotlin plugins in the classpath (#3984)
This commit changes the order of adding atomicfu and kotlin gradle plugin to the classpath.

When `kotlin-gradle-plugin` is added first, `atomicfu-gradle-plugin` is not applied at all and this error is thrown: "e: Module kotlinx.atomicfu cannot be found in the module graph". The change of order fixes the issue. Should be properly investigated.

This behaviour started to reproduce after this change in kotlinx-atomicfu (https://github.com/Kotlin/kotlinx-atomicfu/pull/377).

Here is the corresponding issue in atomicfu: https://github.com/Kotlin/kotlinx-atomicfu/issues/384