Work around #3968, start running integration tests for Native (#3996)
As there were no native targets in smokeTest, the problem with the missing atomicfu dependency (KT-64111) was not addressed in time. The bug was revealed after  release of kotlinx.coroutines with atomicfu-gradle-plugin 0.23.1. Coroutines enabled Native IR transformation, and for this mode, atomicfu-gradle-plugin only provided compile dependency to atomicfu, though implementation dependency was still necessary.

This commit adds the smoke test for Native and temporarily adds an explicit dependency on atomicfu for Native artifacts.

Fixes #3968

Co-authored-by: Maria Sokolova <maria.sokolova@jetbrains.com>