Improve test robustness (#4369)
* Fix testFutureIsDoneAfterChildrenCompleted flakiness

`invokeOnCompletion` can run in parallel to the code that has
already detected a coroutine's completion.
This has happened on the CI once.
The assumption that this can't happen has been removed from the
test.

* Improve println() detection in tests

Now, the printed messages are also preserved, and also, failing to
remove a println capturer from a test no longer affects the
unrelated tests.

* Move robust test cleanup