Use providers.gradle property instead of rootProject.properties in the build scripts (#4421)
See <https://github.com/Kotlin/kotlinx.coroutines/pull/4392/files#r2016110128>
for an explanation of the benefits of this.

Also, fix a bug in `integration-testing` that used to always
treat the build as one with snapshot dependencies, given that
it was always using the snapshot version of `kotlinx.coroutines`
by design.