Remove flaky check from `StateFlowStressTest`.
There's no guarantee that any particular collector will ever encounter
a value emitted by a particular emitter. It's entirely possible for the
value to be overwritten by a different emitter before the collector gets
a chance to collect it. It's very, very unlikely for a collector to miss
the second half of all values emitted by a particular emitter, but it
is possible to happen and this causes the test to be flaky.