Fix a flaky check and a TSan violation in `StateFlowStressTest`. (#4483)
There's no guarantee that any particular collector will ever encounter
a value emitted by a particular emitter. It's entirely possible for the
value to be overwritten by a different emitter before the collector gets
a chance to collect it. It's very unlikely for a collector to miss the
second half of all values emitted by a particular emitter, but it is
still possible and this causes the test to be flaky.

We can instead check if the collector has collected a recent enough
value from *any* emitter. This should be sufficient to verify that the
collector was still running near the end of the test.

Also fixed a race condition in the test when printing test progress.
The race condition is benign, but it causes TSan to fail for the test,
which could prevent it from finding other concurrency bugs.