Remove the assumed pitfall from `consumeEach` documentation
The way to trigger the pitfall that used to be described can be
boiled down to this pattern:
```kotlin
run {
    channel.consumeEach {
        if (channel.isEmpty) {
            // do something
            return@run
        } else {
            // do something else
        }
    }
}
```

However, here, `isEmpty` is already introducing a race condition,
so `consumeEach` itself does not cause any additional issues.

This does not seem like a pitfall in realistic scenarios after all.
`consumeEach` can perform an early return and thus erase
the elements present in the channel, but this also goes for
elements that possibly entered the channel long ago; without
explicitly checking if the channel is empty, there is no way to
distinguish these two scenarios.