Add the EXACTLY_ONCE contract to `suspendCancellableContinuation` (#4574)
`suspendCancellableContinuation` invokes its block exactly once.
`suspendCoroutineUninterceptedOrReturn`,
which is used for implementing `suspendCancellableContinuation`,
already has a contract stating just that,
so for consistency and completeness, we add a contract to
`suspendCancellableContinuation` itself as well.

This will let the compiler recognize that `val`-variables can be
safely assigned inside the lambda, as they won't be reassigned
or left uninitialized (which would be the case if the lambda
executed more than once or possibly not execute at all).

There were no use cases actually reported for this.