Fix `SharedFlow` entering an invalid state
Fixes #4583

See the `SharedFlow.testCancellingSubscriberAndEmitterWithNoBuffer`
test for the problematic scenario, which may arise when
the next emitter and the last collector get cancelled simultaneously.

After the first iteration of the test,
the internal state of the `SharedFlow` was:

* buffer = [null, NO_VALUE]
* replayIndex = 2
* minCollectorIndex = 1
* bufferSize = 1
* queueSize = 0
* nCollectors = 0
* nextIndex = 0

This is clearly invalid, as `bufferSize` can not be anything but 0
with the default `MutableSharedFlow` configuration.

The problem was within the logic that handles unsubscription from
a `SharedFlow` without a buffer if the first emitter is also
cancelled.

There were two separate checks for the cancellation case--
importantly, in that order:
- If the last collector was cancelled,
  the part of the buffer outside the replay buffer was erased,
  since no one was going to consume it.
  This was achieved by ordering the start of the buffer to be
  at *right buffer boundary*.
- If the next emitter was cancelled *and* the buffer is disabled,
  the *right boundary* of the empty buffer was moved past the
  cancelled emitter.

Taken together, these two checks conspired to break the invariants:
if *both* conditions are true, then first, the buffer was set to be
empty, and then, its right boundary was shifted, resulting in a
non-empty buffer when it should be impossible logically.